Moved ProcessUnion.py into Processor.py
